Add optional prefix for commit status notification
This commit aims to extend the functionality of commit status
notifications. It adds an optional key `commitStatusPrefix` to the
Provider spec so that notification controllers in different
environments don't overwrite a commit status for a particular resource.

Example use-case: a single object (kustomization/k8s-gitops) is deployed to
multiple clusters. Specifying this (e.g. to the name of the cluster)
ensures that the notification controllers on each cluster don't
overwrite each others' commit statuses when reporting on
kustomization/k8s-gitops.

Signed-off-by: bcarrier <bcarrier@actbluetech.com>