fix eventhandler test sideeffects with fixture
Introduce event fixtures in the eventhandler test to prevent
sideeffects from the previous test case. The event fixtures are used
to reset the event used in the test cases to an initial fixed value.
The reset is moved to the beginning of subtest to ensure those variables
always get reset at the beginning of a subtest. The previous
implementation in which the reset was done at the end doesn't get
executed when a test fails.
This helps ensure that two consecutive test cases don't use the same
event value.

Signed-off-by: Sunny <darkowlzz@protonmail.com>